implemented the error validation through the service, serviceImplementation and (curso) controller

+++ src/main/java/com/example/escola/DTO/CursoRequestDTO.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class CursoRequestDTO {
     private String nome;
     private Integer cargaHoraria;
-    private Integer categoriaCursoId;
+    private Long categoriaCursoId;
 }
