+++ src/main/java/com/example/escola/domains/CategoriaCurso.java
@@ -13,13 +13,13 @@ import jakarta.persistence.OneToMany;
 public class CategoriaCurso {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String nome;
     
     @OneToMany(mappedBy = "categoriaCurso")
     private List<Curso> cursos;
 
-    public CategoriaCurso(Integer id, String nome) {
+    public CategoriaCurso(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
@@ -27,11 +27,11 @@ public class CategoriaCurso {
     public CategoriaCurso() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
