+++ src/main/java/com/example/escola/repositories/CategoriaCursoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.escola.domains.CategoriaCurso;
 
-public interface CategoriaCursoRepository extends JpaRepository<CategoriaCurso, Integer> {
+public interface CategoriaCursoRepository extends JpaRepository<CategoriaCurso, Long> {
     
 }
